[chore] rename architecture file name
